fix: code fixes
